LayoutTests/imported/w3c/web-platform-tests/editing/other/edit-in-textcontrol-immediately-after-hidden.tentative_editor=input&hide-target=editor-expected.txt


PASS Wait for load event
PASS execCommand("insertText", false, "typed value") in <input>
PASS execCommand("delete") in <input>
PASS <input> is hidden by "keydown" event listener
PASS <input> is hidden by "keypress" event listener
PASS <input> is hidden by "beforeinput" event listener
